Refactored folder name DAO to data_access to match main branch
